requirements.txt
aiodns==1.1.1
aiohttp==3.6.2
apipkg==1.5
appdirs==1.4.4
asn1crypto==1.4.0
asttokens==2.0.4
async-timeout==3.0.1
attrs==20.3.0
base58==2.1.0
bech32==1.2.0
bip32==0.0.8
bitarray==1.2.2
black==20.8b1
bottle==0.12.18
ccxt==1.35.27
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
coincurve==13.0.0
cryptography==3.3.1
cycler==0.10.0
cytoolz==0.11.0
dacite==1.5.1
distlib==0.3.1
ecdsa==0.16.0
erdpy==0.9.8
eth-abi==2.1.1
eth-account==0.5.4
eth-brownie==1.14.6
eth-event==1.2.3
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.8.0
filelock==3.0.12
Flask==1.1.2
gevent==20.12.1
greenlet==1.0.0
hexbytes==0.2.1
hypothesis==6.10.0
iconsdk==1.3.2
idna==2.10
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.7.0a1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
keyboard==0.13.5
kiwisolver==1.3.1
lru-dict==1.1.7
MarkupSafe==1.1.1
matplotlib==3.3.3
mnemonic==0.19
multiaddr==0.0.9
multidict==4.7.6
multipledispatch==0.6.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
numpy==1.19.4
packaging==20.9
parsimonious==0.8.1
pathspec==0.8.1
Pillow==8.0.1
pluggy==0.13.1
prettytable==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
py==1.10.0
py-solc-ast==1.2.8
py-solc-x==1.1.0
pyband==0.0.8
pycares==3.1.1
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.9
Pygments==2.8.1
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyo==1.0.3
pyobjc==7.0.1
pyobjc-core==7.0.1
pyobjc-framework-Accounts==7.0.1
pyobjc-framework-AddressBook==7.0.1
pyobjc-framework-AdSupport==7.0.1
pyobjc-framework-AppleScriptKit==7.0.1
pyobjc-framework-AppleScriptObjC==7.0.1
pyobjc-framework-ApplicationServices==7.0.1
pyobjc-framework-AuthenticationServices==7.0.1
pyobjc-framework-AutomaticAssessmentConfiguration==7.0.1
pyobjc-framework-Automator==7.0.1
pyobjc-framework-AVFoundation==7.0.1
pyobjc-framework-AVKit==7.0.1
pyobjc-framework-BusinessChat==7.0.1
pyobjc-framework-CalendarStore==7.0.1
pyobjc-framework-CFNetwork==7.0.1
pyobjc-framework-CloudKit==7.0.1
pyobjc-framework-Cocoa==7.0.1
pyobjc-framework-Collaboration==7.0.1
pyobjc-framework-ColorSync==7.0.1
pyobjc-framework-Contacts==7.0.1
pyobjc-framework-ContactsUI==7.0.1
pyobjc-framework-CoreAudio==7.0.1
pyobjc-framework-CoreAudioKit==7.0.1
pyobjc-framework-CoreBluetooth==7.0.1
pyobjc-framework-CoreData==7.0.1
pyobjc-framework-CoreHaptics==7.0.1
pyobjc-framework-CoreLocation==7.0.1
pyobjc-framework-CoreMedia==7.0.1
pyobjc-framework-CoreMediaIO==7.0.1
pyobjc-framework-CoreMIDI==7.0.1
pyobjc-framework-CoreML==7.0.1
pyobjc-framework-CoreMotion==7.0.1
pyobjc-framework-CoreServices==7.0.1
pyobjc-framework-CoreSpotlight==7.0.1
pyobjc-framework-CoreText==7.0.1
pyobjc-framework-CoreWLAN==7.0.1
pyobjc-framework-CryptoTokenKit==7.0.1
pyobjc-framework-DeviceCheck==7.0.1
pyobjc-framework-DictionaryServices==7.0.1
pyobjc-framework-DiscRecording==7.0.1
pyobjc-framework-DiscRecordingUI==7.0.1
pyobjc-framework-DiskArbitration==7.0.1
pyobjc-framework-DVDPlayback==7.0.1
pyobjc-framework-EventKit==7.0.1
pyobjc-framework-ExceptionHandling==7.0.1
pyobjc-framework-ExecutionPolicy==7.0.1
pyobjc-framework-ExternalAccessory==7.0.1
pyobjc-framework-FileProvider==7.0.1
pyobjc-framework-FileProviderUI==7.0.1
pyobjc-framework-FinderSync==7.0.1
pyobjc-framework-FSEvents==7.0.1
pyobjc-framework-GameCenter==7.0.1
pyobjc-framework-GameController==7.0.1
pyobjc-framework-GameKit==7.0.1
pyobjc-framework-GameplayKit==7.0.1
pyobjc-framework-ImageCaptureCore==7.0.1
pyobjc-framework-IMServicePlugIn==7.0.1
pyobjc-framework-InputMethodKit==7.0.1
pyobjc-framework-InstallerPlugins==7.0.1
pyobjc-framework-InstantMessage==7.0.1
pyobjc-framework-Intents==7.0.1
pyobjc-framework-IOSurface==7.0.1
pyobjc-framework-iTunesLibrary==7.0.1
pyobjc-framework-LatentSemanticMapping==7.0.1
pyobjc-framework-LaunchServices==7.0.1
pyobjc-framework-libdispatch==7.0.1
pyobjc-framework-LinkPresentation==7.0.1
pyobjc-framework-LocalAuthentication==7.0.1
pyobjc-framework-MapKit==7.0.1
pyobjc-framework-MediaAccessibility==7.0.1
pyobjc-framework-MediaLibrary==7.0.1
pyobjc-framework-MediaPlayer==7.0.1
pyobjc-framework-MediaToolbox==7.0.1
pyobjc-framework-Metal==7.0.1
pyobjc-framework-MetalKit==7.0.1
pyobjc-framework-MetalPerformanceShaders==7.0.1
pyobjc-framework-ModelIO==7.0.1
pyobjc-framework-MultipeerConnectivity==7.0.1
pyobjc-framework-NaturalLanguage==7.0.1
pyobjc-framework-NetFS==7.0.1
pyobjc-framework-Network==7.0.1
pyobjc-framework-NetworkExtension==7.0.1
pyobjc-framework-NotificationCenter==7.0.1
pyobjc-framework-OpenDirectory==7.0.1
pyobjc-framework-OSAKit==7.0.1
pyobjc-framework-OSLog==7.0.1
pyobjc-framework-PencilKit==7.0.1
pyobjc-framework-Photos==7.0.1
pyobjc-framework-PhotosUI==7.0.1
pyobjc-framework-PreferencePanes==7.0.1
pyobjc-framework-PushKit==7.0.1
pyobjc-framework-Quartz==7.0.1
pyobjc-framework-QuickLookThumbnailing==7.0.1
pyobjc-framework-SafariServices==7.0.1
pyobjc-framework-SceneKit==7.0.1
pyobjc-framework-ScreenSaver==7.0.1
pyobjc-framework-ScriptingBridge==7.0.1
pyobjc-framework-SearchKit==7.0.1
pyobjc-framework-Security==7.0.1
pyobjc-framework-SecurityFoundation==7.0.1
pyobjc-framework-SecurityInterface==7.0.1
pyobjc-framework-ServiceManagement==7.0.1
pyobjc-framework-Social==7.0.1
pyobjc-framework-SoundAnalysis==7.0.1
pyobjc-framework-Speech==7.0.1
pyobjc-framework-SpriteKit==7.0.1
pyobjc-framework-StoreKit==7.0.1
pyobjc-framework-SyncServices==7.0.1
pyobjc-framework-SystemConfiguration==7.0.1
pyobjc-framework-SystemExtensions==7.0.1
pyobjc-framework-UserNotifications==7.0.1
pyobjc-framework-VideoSubscriberAccount==7.0.1
pyobjc-framework-VideoToolbox==7.0.1
pyobjc-framework-Vision==7.0.1
pyobjc-framework-WebKit==7.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.6.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rlp==1.2.0
semantic-version==2.8.5
sha3==0.2.1
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.2
toolz==0.11.1
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
varint==1.0.2
virtualenv==20.2.2
vvm==0.1.0
vyper==0.2.12
wcwidth==0.2.5
web3==5.18.0
websocket==0.2.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.1.0
zope.event==4.5.0
zope.interface==5.2.0
